// src/api/generated/models/ConversationHelpRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ConversationHelpRequest = {
    ai_response: string;
    conversation_context: Array<Record<string, string>>;
    target_language: string;
    user_language: string;
    proficiency_level: string;
    topic?: (string | null);
};