// src/api/generated/models/UserHelpSettings.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type UserHelpSettings = {
    user_id: string;
    help_enabled?: boolean;
    help_language?: string;
    show_pronunciation?: boolean;
    show_grammar_tips?: boolean;
    show_cultural_notes?: boolean;
    show_vocabulary?: boolean;
    created_at?: string;
    updated_at?: string;
};